fix: Move wallet info to and balances to event api
